MLB Scraper - Closing Lines and ALL Line Movement
Input your year, start month, and end month and have an output of 2 databases:
1. all closing lines for Full-Game Moneyline, Runline, Total + First-Five Inning Moneyline, Runline, and Total
2. all line changes made from open to close
Based on the amount of data that needs to be pulled, each day takes ~3-5 minutes to collect all necessary info. Let it run for a while. This script will also return a file that shows all gamedays that did not scrape completely. 

Note: this script pulls for 5 books, Pinnacle, 5dimes, Heritage, Bovada, and BetOnline. It can be changed to include other books as well